+++ common-util/FunnelLandingPage/6AssetManagement/index.jsx
@@ -18,10 +18,10 @@ import Balancer from './Balancer';
 import CallToActionButton from '../Hero/CallToActionButton';
 
 // part of the homepage
-import WhatIsAutonolas from '../../../components/HomePage/2WhatIsAutonolas';
-import DecentralizedAndSophisticated from '../../../components/HomePage/3DecentralizedAndSophisticated';
-import HowDoAutonolasWork from '../../../components/HomePage/4HowDoAutonolasWork';
-import QuickIntroArticles from '../../../components/HomePage/5QuickIntroArticles';
+import WhatIsAutonolas from '../../../components/HomePage/Learn/2WhatIsAutonolas';
+import DecentralizedAndSophisticated from '../../../components/HomePage/Learn/3DecentralizedAndSophisticated';
+import HowDoAutonolasWork from '../../../components/HomePage/Learn/4HowDoAutonolasWork';
+import QuickIntroArticles from '../../../components/HomePage/Learn/5QuickIntroArticles';
 import { LearnContainer } from '../../../components/HomePage/styles';
 
 import { dividerCss } from '../styles';
