+++ common-util/FunnelLandingPage/7AutonomousEcosystem /index.jsx
@@ -8,10 +8,10 @@ import { HeaderText } from 'common-util/styles';
 import { BENEFITS_LIST, WHAT_IT_INVOLVES, EXISTING_SERVICE } from './data';
 
 // part of the homepage
-import WhatIsAutonolas from '../../../components/HomePage/2WhatIsAutonolas';
-import DecentralizedAndSophisticated from '../../../components/HomePage/3DecentralizedAndSophisticated';
-import HowDoAutonolasWork from '../../../components/HomePage/4HowDoAutonolasWork';
-import QuickIntroArticles from '../../../components/HomePage/5QuickIntroArticles';
+import WhatIsAutonolas from '../../../components/HomePage/Learn/2WhatIsAutonolas';
+import DecentralizedAndSophisticated from '../../../components/HomePage/Learn/3DecentralizedAndSophisticated';
+import HowDoAutonolasWork from '../../../components/HomePage/Learn/4HowDoAutonolasWork';
+import QuickIntroArticles from '../../../components/HomePage/Learn/5QuickIntroArticles';
 import { LearnContainer } from '../../../components/HomePage/styles';
 
 import CallToActionButton from '../Hero/CallToActionButton';
